+++ node.cpp
@@ -8,7 +8,7 @@ void Latch::set_source(Source* s) {
   source = s;
 }
 
-int16_t Latch::next(int64_t tick, int32_t phi) {
+int16_t Latch::next(int32_t tick, int32_t phi) {
   if (! on) {
     On(*this);
     previous = source->next(tick, phi);
@@ -27,6 +27,6 @@ On::~On() {
 
 Constant::Constant(int16_t v) : value(v) {};
 
-int16_t Constant::next(int64_t tick, int32_t phi) {
+int16_t Constant::next(int32_t tick, int32_t phi) {
   return value;
 }
