+++ oscillator.h
@@ -23,7 +23,7 @@ public:
 
   Oscillator(Wavetable& wave, std::unique_ptr<Frequency> freq);
   Oscillator(Wavetable& wave, std::unique_ptr<Frequency> freq, Amplitude amp);
-  int16_t next(int64_t tick, int32_t phi) override;
+  int16_t next(int32_t tick, int32_t phi) override;
   const Frequency& get_frequency();
   
 private:
