+++ source.h
@@ -43,14 +43,14 @@ public:
   
   // tick needs to be signed because phi can be negative and we need
   // to add the two.  also, next is not const because of Latch.
-  virtual int16_t next(int64_t tick, int32_t phi) = 0;
-  int16_t next(int64_t tick);
+  virtual int16_t next(int32_t tick, int32_t phi) = 0;
+  int16_t next(int32_t tick);
   
 };
 
 
-int64_t tick2idx(int64_t tick);
-int64_t hz2tick(float hz);
+size_t tick2idx(int32_t tick);
+int32_t hz2tick(float hz);
 
 
 #endif
